Fix script for Unicode and hack for subdirs
Added explicit decoding/encoding for UTF-8 so script would run
properly on Windows (and really this is a best practice even for
running on Mac OS or Linux).

Added hack to dir iterator - current Hugo docs have some sub-dirs.
For now, these are skipped, but I'll update the code to actually
walk instead of just iterating names in a dir (barely more lines
of code).

Finally, ran on current Hugo docs as of 2018-02-20. It would be
interesting to write a script to run this for each version since
the docs started, that would be a quick way to actually diff all
the docs at once.